requirements.txt
psycopg2-binary
sqlalchemy
pandas
python-dotenv
